Load component and product data on app startup, rounded component price data
