+++ src/main/java/com/example/etl1/controller/ComponentsController.java
@@ -1,19 +1,11 @@
 package com.example.etl1.controller;
 
-import com.example.etl1.model.components.*;
 import com.example.etl1.repository.components.*;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.servlet.ModelAndView;
-
-import java.io.File;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
 
 @Controller
 public class ComponentsController {
@@ -28,12 +20,6 @@ public class ComponentsController {
     CpuCoolerRepository cpuCoolerRepository;
 
     @Autowired
-    FanRpmRepository fanRPMRepository;
-
-    @Autowired
-    FanNoiseLevelRepository fanNoiseLevelRepository;
-
-    @Autowired
     GraphicsCardRepository graphicsCardRepository;
 
     @Autowired
@@ -43,115 +29,10 @@ public class ComponentsController {
     MemoryRepository memoryRepository;
 
     @Autowired
-    MemorySpeedRepository memorySpeedRepository;
-
-    @Autowired
-    MemoryModuleRepository memoryModuleRepository;
-
-    @Autowired
     MotherboardRepository motherboardRepository;
 
     @Autowired
     PowerSupplyRepository powerSupplyRepository;
-
-    @GetMapping("/load-component-data")
-    public String populateComponentsData() throws IOException {
-        caseRepository.deleteAll();
-        cpuRepository.deleteAll();
-        cpuCoolerRepository.deleteAll();
-        graphicsCardRepository.deleteAll();
-        internalStorageRepository.deleteAll();
-        memoryRepository.deleteAll();
-        motherboardRepository.deleteAll();
-        powerSupplyRepository.deleteAll();
-
-        ObjectMapper mapper = new ObjectMapper();
-        Case[] cases = mapper.readValue(new File(System.getProperty("user.dir") + "/src/main/resources/PC_Components_Data/case.json"), Case[].class);
-        Cpu[] cpus = mapper.readValue(new File(System.getProperty("user.dir") + "/src/main/resources/PC_Components_Data/cpu.json"), Cpu[].class);
-        CpuCooler[] cpuCoolers = mapper.readValue(new File(System.getProperty("user.dir") + "/src/main/resources/PC_Components_Data/cpu-cooler.json"), CpuCooler[].class);
-        GraphicsCard[] graphicsCards = mapper.readValue(new File(System.getProperty("user.dir") + "/src/main/resources/PC_Components_Data/graphics-card.json"), GraphicsCard[].class);
-        InternalStorage[] internalStorages = mapper.readValue(new File(System.getProperty("user.dir") + "/src/main/resources/PC_Components_Data/internal-storage-device.json"), InternalStorage[].class);
-        Memory[] memory = mapper.readValue(new File(System.getProperty("user.dir") + "/src/main/resources/PC_Components_Data/memory.json"), Memory[].class);
-        Motherboard[] motherboards = mapper.readValue(new File(System.getProperty("user.dir") + "/src/main/resources/PC_Components_Data/motherboard.json"), Motherboard[].class);
-        PowerSupply[] powerSupplies = mapper.readValue(new File(System.getProperty("user.dir") + "/src/main/resources/PC_Components_Data/power-supply.json"), PowerSupply[].class);
-
-        convertComponentPriceToPounds(cases);
-        convertComponentPriceToPounds(cpus);
-        convertComponentPriceToPounds(cpuCoolers);
-        convertComponentPriceToPounds(graphicsCards);
-        convertComponentPriceToPounds(internalStorages);
-        convertComponentPriceToPounds(memory);
-        convertComponentPriceToPounds(motherboards);
-        convertComponentPriceToPounds(powerSupplies);
-
-        caseRepository.saveAll(Arrays.asList(cases));
-        cpuRepository.saveAll(Arrays.asList(cpus));
-
-        cpuCoolerRepository.saveAll(Arrays.asList(cpuCoolers));
-
-        List<FanRpm> allRPMs = new ArrayList<>();
-        List<FanNoiseLevel> allNoiseLevels = new ArrayList<>();
-
-        for (CpuCooler cpuCooler : cpuCoolers) {
-            List<FanRpm> rpms = cpuCooler.getRpm();
-            if (!rpms.isEmpty()) {
-                for (FanRpm rpm : rpms) {
-                    if (rpm != null) {
-                        rpm.setCpu_cooler(cpuCooler);
-                        allRPMs.add(rpm);
-                    }
-                }
-            }
-            List<FanNoiseLevel> noiseLevels = cpuCooler.getNoise_level();
-            if (!noiseLevels.isEmpty()) {
-                for (FanNoiseLevel noiseLevel : noiseLevels) {
-                    if (noiseLevel != null) {
-                        noiseLevel.setCpu_cooler(cpuCooler);
-                        allNoiseLevels.add(noiseLevel);
-                    }
-                }
-            }
-        }
-
-        fanRPMRepository.saveAll(allRPMs);
-        fanNoiseLevelRepository.saveAll(allNoiseLevels);
-        graphicsCardRepository.saveAll(Arrays.asList(graphicsCards));
-        internalStorageRepository.saveAll(Arrays.asList(internalStorages));
-        memoryRepository.saveAll(Arrays.asList(memory));
-
-        List<MemorySpeed> allMemorySpeeds = new ArrayList<>();
-        List<MemoryModule> allMemoryModules = new ArrayList<>();
-
-        for (Memory mem : memory) {
-            List<MemorySpeed> memorySpeeds = mem.getSpeed();
-            if (!memorySpeeds.isEmpty()) {
-                for (MemorySpeed memorySpeed : memorySpeeds) {
-                    if (memorySpeed != null) {
-                        memorySpeed.setMemory(mem);
-                        allMemorySpeeds.add(memorySpeed);
-                    }
-                }
-            }
-
-            List<MemoryModule> memoryModules = mem.getModules();
-            if (!memoryModules.isEmpty()) {
-                for (MemoryModule memoryModule : memoryModules) {
-                    if (memoryModule != null) {
-                        memoryModule.setMemory(mem);
-                        allMemoryModules.add(memoryModule);
-                    }
-                }
-            }
-        }
-
-        memorySpeedRepository.saveAll(allMemorySpeeds);
-        memoryModuleRepository.saveAll(allMemoryModules);
-
-        motherboardRepository.saveAll(Arrays.asList(motherboards));
-        powerSupplyRepository.saveAll(Arrays.asList(powerSupplies));
-
-        return "redirect:/";
-    }
 
     @GetMapping("/components/cases")
     public ModelAndView viewCases(String sortBy, String order) {
@@ -291,11 +172,5 @@ public class ComponentsController {
         ModelAndView modelAndView = new ModelAndView("/components/cpus");
         modelAndView.addObject("cpus", cpuRepository.findByNameContaining("AMD"));
         return modelAndView;
-    }
-
-    private void convertComponentPriceToPounds(Component[] components) {
-        for (Component component : components) {
-            component.setPrice(component.getPrice() * 0.75);
-        }
     }
 }
